update: rename api folder
