+++ backend/src/server.ts
@@ -1,40 +1,40 @@
-import express from "express";
-import cors from "cors";
-import multer from "multer";
 import bodyParser from "body-parser";
+import cors from "cors";
+import express from "express";
+import multer from "multer";
 import {
   processFile,
   retrieveAndReassembleData,
 } from "./services/storageManager/index.js";
 import {
-  TransactionResult,
   createApi,
   retrieveRemarkFromTransaction,
+  TransactionResult,
 } from "./services/transactionManager/index.js";
 import { isJson } from "./utils/index.js";
 
-import dotenv from "dotenv";
-import { FolderTreeSchema } from "./models/folderTree.js";
-import {
-  processTree,
-  uploadFile,
-  uploadTree,
-} from "./services/storageManager/storageManager.js";
 import {
   cidToString,
   IPLDNodeData,
   OffchainMetadata,
 } from "@autonomys/auto-drive";
 import { decode } from "@ipld/dag-pb";
+import dotenv from "dotenv";
+import { FolderTreeSchema } from "./models/folderTree.js";
 import { NodeWithMetadata } from "./models/nodeWithMetadata.js";
-import { getNode } from "./api/nodes.js";
-import { getMetadata, searchMetadataByCID } from "./api/metadata.js";
+import { transactionResultsRepository } from "./repositories/transactionResults.js";
+import {
+  processTree,
+  uploadFile,
+  uploadTree,
+} from "./services/storageManager/storageManager.js";
+import { uploadManager } from "./services/uploadManager/index.js";
+import { getMetadata, searchMetadataByCID } from "./useCases/metadata.js";
+import { getNode } from "./useCases/nodes.js";
 import {
   getHeadTransactionResults,
   getNodeTransactionResult,
-} from "./api/transactionResults.js";
-import { uploadManager } from "./services/uploadManager/index.js";
-import { transactionResultsRepository } from "./repositories/transactionResults.js";
+} from "./useCases/transactionResults.js";
 
 dotenv.config();
