+++ backend/src/useCases/transactionResults.ts
